+++ Models/PendingUser.cs
@@ -2,13 +2,13 @@ namespace UABackbone_Backend.Models;
 
 public class PendingUser
 {
-    public ushort   Id           { get; set; }
+    public int      Id           { get; set; }
     public string   Username     { get; set; }
     public string   PasswordHash { get; set; }
     public string   FirstName    { get; set; }
     public string   LastName     { get; set; }
     public string   Email        { get; set; }
-    public short    Local        { get; set; }
+    public int      Local        { get; set; }
     public byte[]   UaCardImage  { get; set; }
     public DateTime SubmittedAt  { get; set; }
 }
